incident-generator/requirements.txt
requests
python-dotenv
google-generativeai
Flask
flask-cors
